melanoma-app/requirements.txt
streamlit>=1.36
numpy
Pillow
scikit-learn
opencv-python-headless
torch
torchvision
scipy
pandas
requests
